requirements.txt
alembic==1.0.0
bcrypt==3.1.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
passlib==1.7.1
pkg-resources==0.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.6.1
pycparser==2.19
PyJWT==1.6.4
python-dateutil==2.7.3
python-editor==1.0.3
requests==2.21.0
requirements==0.1
six==1.11.0
SQLAlchemy==1.2.12
urllib3==1.24.1
Werkzeug==0.14.1
